Make list objects submodules, untested
