+++ lists/__init__.py
@@ -1,0 +1,10 @@
+﻿# -*- coding: utf-8 -*-
+#Falcon generic list management
+#Copyright (C) 2019-2020 Yukio Nozawa <personal@nyanchangames.com>
+#Copyright (C) 2019-2020 yamahubuki <itiro.ishino@gmail.com>
+#Note: All comments except these top lines will be written in Japanese. 
+
+from lists.file import *
+from lists.drive import *
+from lists.stream import *
+from lists.constants import *
